+++ dtslint/ts3.5/TaskDecoder.ts
@@ -1,46 +1,96 @@
+import { pipe } from 'fp-ts/lib/pipeable'
+import * as DE from '../../src/DecodeError'
 import * as _ from '../../src/TaskDecoder'
+import * as FS from '../../src/FreeSemigroup'
 
-declare const string: _.TaskDecoder<string, string>
+declare const StringFromString: _.TaskDecoder<string, string>
 declare const NumberFromString: _.TaskDecoder<string, number>
 
 //
-// type
+// TypeOf
+//
+
+// $ExpectType number
+export type TypeOfNumberFromString = _.TypeOf<typeof NumberFromString>
+
+//
+// InputOf
+//
+
+// $ExpectType string
+export type InputOfNumberFromString = _.InputOf<typeof NumberFromString>
+
+//
+// ktype
 //
 
 // $ExpectType TaskDecoder<{ a: string; }, { a: number; }>
-_.type({
+_.ktype({
   a: NumberFromString
 })
 
 //
-// partial
+// kpartial
 //
 
 // $ExpectType TaskDecoder<{ a: string; }, Partial<{ a: number; }>>
-_.partial({
+_.kpartial({
   a: NumberFromString
 })
 
 //
-// tuple
+// ktuple
 //
 
 // $ExpectType TaskDecoder<[string], [number]>
-_.tuple(NumberFromString)
+_.ktuple(NumberFromString)
 
 //
-// sum
+// ksum
 //
 
 // $ExpectType TaskDecoder<{ _tag: unknown; a: string; } | { _tag: unknown; b: string; }, { _tag: "A"; a: number; } | { _tag: "B"; b: number; }>
-_.sum('_tag')({
-  A: _.type({ _tag: _.literal('A'), a: NumberFromString }),
-  B: _.type({ _tag: _.literal('B'), b: NumberFromString })
+_.ksum('_tag')({
+  A: _.ktype({ _tag: _.literal('A'), a: NumberFromString }),
+  B: _.ktype({ _tag: _.literal('B'), b: NumberFromString })
 })
 
 //
 // union
 //
 
+// $ExpectType TaskDecoder<unknown, string | number>
+_.union(_.number, _.string)
+
 // $ExpectType TaskDecoder<string, string | number>
-_.union(NumberFromString, string)
+_.union(NumberFromString, StringFromString)
+
+// -------------------------------------------------------------------------------------
+// unknown input
+// -------------------------------------------------------------------------------------
+
+// $ExpectType TaskDecoder<unknown, { a: string; b: { c: number; }; }>
+const A = _.type({
+  a: _.string,
+  b: _.type({
+    c: _.number
+  })
+})
+
+// $ExpectType TaskDecoder<unknown, Partial<{ a: string; b: Partial<{ c: number; }>; }>>
+_.partial({
+  a: _.string,
+  b: _.partial({
+    c: _.number
+  })
+})
+
+//
+// mapLeftWithInput
+//
+
+// $ExpectType TaskDecoder<unknown, number>
+pipe(
+  _.number,
+  _.mapLeftWithInput((u) => FS.of(DE.leaf(u, 'not a number')))
+)